feat: 투자 성향 파악 겡게임 타입 오류 수정

+++ back/src/main/java/com/stock/survive/repository/tendency/TendencyGameSessionRepository.java
@@ -14,9 +14,9 @@ import java.util.Optional;
 @Repository
 public interface TendencyGameSessionRepository extends JpaRepository<TendencyGameSession, Long> {
 
-    Optional<TendencyGameSession> findByIdAndUser_Id(Long id, Integer userId);
+    Optional<TendencyGameSession> findByIdAndUser_Id(Long id, Long userId);
 
-    List<TendencyGameSession> findByUser_IdOrderByStartedAtDesc(Integer userId);
+    List<TendencyGameSession> findByUser_IdOrderByStartedAtDesc(Long userId);
     
     List<TendencyGameSession> findByTickerAndStartedAtBetween(String ticker, LocalDateTime startDate, LocalDateTime endDate);
 }
